ux(sidebar): move New Location out of toolbar into Settings (#131)
Beta testers consistently misread the sidebar's primary toolbar `+`
as "add item" rather than "add location" — locations are a rare-
action surface that doesn't deserve top-level real estate. This PR
moves location create / edit / delete into the Settings screen
(alongside household management, which already lives there) and
extracts a `LocationsSection` SwiftUI view to keep the wiring
testable.

Sidebar changes:
- Drops the primary `+` ToolbarItem and its location-form sheet.
- Drops the row-level swipe edit/delete actions per the issue's
  recommendation, so all location mutations live in one place.
- Empty-state copy now points at the gear ("tap the gear to add
  one") instead of the (removed) `+`.
- Keeps `RemoteChangeMonitor.changeToken`-driven cross-device
  reload; adds an `onChange(of: isPresentingSettings)` reload edge
  so same-device Settings edits reflect after dismiss (the
  changeToken path skips local-author writes).
- Reconciles selection: if the active location is deleted from
  Settings, snap back to All Items.

Settings changes:
- New `LocationsSection` between Household and Expiry Reminders.
- Reuses existing `LocationFormView` unchanged.
- Same destructive-confirmation copy ("This also removes every
  item inside it.") as the old sidebar dialog.

NOTE for review: this PR alone leaves the sidebar with no primary
toolbar action — that's the documented in-between state until #132
(sibling) repurposes the freed slot as "New Item." Hold merging
until #132 is ready, or merge both back-to-back before any
TestFlight push.
